Hotfix: addresses #3983
The combination of groupByResponseName and groupByParentType is overzealous
in combining fields. In the case of "duplicate" field selections, multiple
fieldsForParentType may exist, but only the first is ever "looked at" for
decision making. In the case that the first one will be ignored (has no
possibleTypes), the second one was also thrown out. The root of this problem
indicates that the grouping itself is incorrect and we should reconsider
the grouping strategy.

By filtering out "invalid" (no possibleTypes) fields ahead of time, this
issue is worked around. As noted in the large comment associated with this
commit, I don't trust the fragility of this approach and believe that it
should be addressed. This requires a bit more thought on my part. In the
mean time, this is a sensible fix that I'd prefer not to withhold since
it's an undeniably an improvement on the current situation.